Remove separate fragment parsers and put everything in the same grammar
