+++ src/main/antlr4/v1_0/WdlV1Parser.g4
@@ -278,5 +278,15 @@ document_element
 	;
 
 document
-	: version document_element* (workflow document_element*)? EOF
+	: version document_element* (workflow document_element*)?
 	;
+
+document_or_fragment
+  : document
+  | wdl_type
+  | expr
+  ;
+
+top
+  : document_or_fragment EOF
+  ;
