+++ src/main/scala/wdlTools/syntax/Parsers.scala
@@ -14,13 +14,9 @@ case class Parsers(opts: Options = Options(), defaultLoader: Option[SourceCode.L
       WdlVersion.Draft_2 -> draft_2.ParseAll(opts, loader),
       WdlVersion.V1_0 -> v1_0.ParseAll(opts, loader)
   )
-  private lazy val typeParsers: Map[WdlVersion, WdlTypeParser] = Map(
-      WdlVersion.Draft_2 -> draft_2.ParseType(opts),
-      WdlVersion.V1_0 -> v1_0.ParseType(opts)
-  )
-  private lazy val exprParsers: Map[WdlVersion, WdlExprParser] = Map(
-      WdlVersion.Draft_2 -> draft_2.ParseExpr(opts),
-      WdlVersion.V1_0 -> v1_0.ParseExpr(opts)
+  private lazy val fragmentParsers: Map[WdlVersion, WdlFragmentParser] = Map(
+      WdlVersion.Draft_2 -> draft_2.ParseFragment(opts),
+      WdlVersion.V1_0 -> v1_0.ParseFragment(opts)
   )
 
   def getParser(url: URL): WdlParser = {
@@ -60,11 +56,7 @@ case class Parsers(opts: Options = Options(), defaultLoader: Option[SourceCode.L
     parser.Walker(url, Some(sourceCode), results)
   }
 
-  def getTypeParser(wdlVersion: WdlVersion): WdlTypeParser = {
-    typeParsers(wdlVersion)
-  }
-
-  def getExprParser(wdlVersion: WdlVersion): WdlExprParser = {
-    exprParsers(wdlVersion)
+  def getFragmentParser(wdlVersion: WdlVersion): WdlFragmentParser = {
+    fragmentParsers(wdlVersion)
   }
 }
