+++ src/test/scala/wdlTools/syntax/v1_0/ConcreteSyntaxTest.scala
@@ -19,6 +19,7 @@ class ConcreteSyntaxTest extends FlatSpec with Matchers {
       localDirectories = Some(Vector(tasksDir, workflowsDir, structsDir))
   )
   private val loader = SourceCode.Loader(opts)
+  private val grammarFactory = WdlV1GrammarFactory(opts)
 
   private def getTaskSource(fname: String): SourceCode = {
     loader.apply(Util.getURL(tasksDir.resolve(fname)))
@@ -32,8 +33,12 @@ class ConcreteSyntaxTest extends FlatSpec with Matchers {
     loader.apply(Util.getURL(structsDir.resolve(fname)))
   }
 
+  private def getDocument(sourceCode: SourceCode, conf: Options = opts): Document = {
+    ParseOne(conf, grammarFactory.createGrammar(sourceCode.toString), Some(sourceCode.url)).parseDocument
+  }
+
   it should "handle various types" in {
-    val doc = ParseDocument.apply(getTaskSource("types.wdl"), opts)
+    val doc = getDocument(getTaskSource("types.wdl"))
 
     doc.elements.size shouldBe 1
     val elem = doc.elements(0)
@@ -82,7 +87,7 @@ class ConcreteSyntaxTest extends FlatSpec with Matchers {
   }
 
   it should "handle types and expressions" in {
-    val doc = ParseDocument.apply(getTaskSource("expressions.wdl"), opts)
+    val doc = getDocument(getTaskSource("expressions.wdl"))
 
     doc.version.value shouldBe WdlVersion.V1_0
     doc.elements.size shouldBe 1
@@ -258,7 +263,7 @@ class ConcreteSyntaxTest extends FlatSpec with Matchers {
   }
 
   it should "handle get name" in {
-    val doc = ParseDocument.apply(getTaskSource("get_name_bug.wdl"), opts)
+    val doc = getDocument(getTaskSource("get_name_bug.wdl"))
 
     doc.version.value shouldBe WdlVersion.V1_0
     doc.elements.size shouldBe 1
@@ -282,12 +287,12 @@ class ConcreteSyntaxTest extends FlatSpec with Matchers {
 
   it should "detect a wrong comment style" in {
     assertThrows[Exception] {
-      ParseDocument.apply(getTaskSource("wrong_comment_style.wdl"), opts)
+      getDocument(getTaskSource("wrong_comment_style.wdl"))
     }
   }
 
   it should "parse a task with an output section only" in {
-    val doc = ParseDocument.apply(getTaskSource("output_section.wdl"), opts)
+    val doc = getDocument(getTaskSource("output_section.wdl"))
 
     doc.version.value shouldBe WdlVersion.V1_0
     doc.elements.size shouldBe 1
@@ -304,7 +309,7 @@ class ConcreteSyntaxTest extends FlatSpec with Matchers {
   }
 
   it should "parse a task" in {
-    val doc = ParseDocument.apply(getTaskSource("wc.wdl"), opts)
+    val doc = getDocument(getTaskSource("wc.wdl"))
 
     doc.version.value shouldBe WdlVersion.V1_0
     doc.elements.size shouldBe 1
@@ -351,12 +356,12 @@ class ConcreteSyntaxTest extends FlatSpec with Matchers {
 
   it should "detect when a task section appears twice" in {
     assertThrows[Exception] {
-      ParseDocument.apply(getTaskSource("multiple_input_section.wdl"), opts)
+      getDocument(getTaskSource("multiple_input_section.wdl"))
     }
   }
 
   it should "handle string interpolation" in {
-    val doc = ParseDocument.apply(getTaskSource("interpolation.wdl"), opts)
+    val doc = getDocument(getTaskSource("interpolation.wdl"))
 
     doc.version.value shouldBe WdlVersion.V1_0
     doc.elements.size shouldBe 1
@@ -384,7 +389,7 @@ class ConcreteSyntaxTest extends FlatSpec with Matchers {
   }
 
   it should "parse structs" in {
-    val doc = ParseDocument.apply(getStructSource("I.wdl"), opts)
+    val doc = getDocument(getStructSource("I.wdl"))
 
     doc.version.value shouldBe WdlVersion.V1_0
     val structs = doc.elements.collect {
@@ -408,7 +413,7 @@ class ConcreteSyntaxTest extends FlatSpec with Matchers {
   }
 
   it should "parse a simple workflow" taggedAs Edge in {
-    val doc = ParseDocument.apply(getWorkflowSource("I.wdl"), opts)
+    val doc = getDocument(getWorkflowSource("I.wdl"))
     doc.elements.size shouldBe 0
 
     doc.version.value shouldBe WdlVersion.V1_0
@@ -465,7 +470,7 @@ class ConcreteSyntaxTest extends FlatSpec with Matchers {
   }
 
   it should "handle import statements" in {
-    val doc = ParseDocument.apply(getWorkflowSource("imports.wdl"), opts)
+    val doc = getDocument(getWorkflowSource("imports.wdl"))
 
     doc.version.value shouldBe WdlVersion.V1_0
 
@@ -486,12 +491,12 @@ class ConcreteSyntaxTest extends FlatSpec with Matchers {
 
   it should "correctly report an error" in {
     assertThrows[Exception] {
-      val _ = ParseDocument.apply(getWorkflowSource("bad_declaration.wdl"), opts)
+      val _ = getDocument(getWorkflowSource("bad_declaration.wdl"))
     }
   }
 
   it should "handle chained operations" taggedAs Edge in {
-    val doc = ParseDocument.apply(getTaskSource("bug16-chained-operations.wdl"), opts)
+    val doc = getDocument(getTaskSource("bug16-chained-operations.wdl"))
 
     doc.elements.size shouldBe 1
     val elem = doc.elements(0)
@@ -509,7 +514,7 @@ class ConcreteSyntaxTest extends FlatSpec with Matchers {
   }
 
   it should "handle chained operations in a workflow" in {
-    val doc = ParseDocument.apply(getWorkflowSource("chained_expr.wdl"), opts)
+    val doc = getDocument(getWorkflowSource("chained_expr.wdl"))
     doc.elements.size shouldBe 0
 
     doc.version.value shouldBe WdlVersion.V1_0
@@ -528,7 +533,7 @@ class ConcreteSyntaxTest extends FlatSpec with Matchers {
     val url =
       "https://raw.githubusercontent.com/gatk-workflows/gatk4-germline-snps-indels/master/tasks/JointGenotypingTasks-terra.wdl"
     val sourceCode = loader.apply(Util.getURL(url))
-    val doc = ParseDocument.apply(sourceCode, opts)
+    val doc = getDocument(sourceCode)
 
     doc.version.value shouldBe WdlVersion.V1_0
     doc.elements.size shouldBe 19
